// types/types.ts
export interface ILead {
  id?: number;
  source: null | string;
  owner: null | string;
  name: null | string;
  source_link: null | string;
  courier: 'Yes' | 'No';
  mobile_number: null | string;
  email: null | string;
  province: null | string;
  town: null | string;
  exact_location: null | string;
  pipeline_stage: EPipelineStage;
  deal_value: string;
  probability: string;
  expected_revenue: string;
  follow_up_date: Date;
  close_date: null | Date;
  status: ELeadStatus;
  comments: null | string;
}

export enum EPipelineStage {
  Qualification = 'Qualification',
  ProposalQuotation = 'Proposal|Quotation',
  Negotiation = 'Negotiation',
  PendingApproval = 'Pending Approval',
  ClosedWon = 'Closed-Won',
  ClosedLost = 'Closed-Lost'
}

export enum ELeadStatus {
  Open = 'Open',
  InProgress = 'In Progress',
  ClosedWon = 'Closed-Won',
  ClosedLost = 'Closed-Lost',
}